Added first functionality for web extension. Removed own web-server.
