--- index.html
@@ -1,29 +1,0 @@
-<!DOCTYPE html>
-<html>
-  <head>
-    <meta charset="utf-8" />
-    <title>ECharts</title>
-    <!-- Include the ECharts file you just downloaded -->
-    <script src="echarts.js"></script>
-  </head>
-  <body>
-    <!-- Prepare a DOM with a defined width and height for ECharts -->
-    <div id="main" ></div>
-    <script>
-      var jsopts = { solution: 42 };
-      var myChart = echarts.init(document.getElementById('main'),null,{
-        width: window.innerWidth-10,
-        height: window.innerHeight-20,
-      });
-
-      window.addEventListener('resize', function() {
-        myChart.resize({
-          width: window.innerWidth-10,
-          height: window.innerHeight-20,
-        });
-      });
-
-      myChart.setOption(jsopts);
-    </script>
-  </body>
-</html>
