+++ lib/eslint.config.mjs
@@ -1,0 +1,7 @@
+// eslint.config.js
+// See "Globally ignoring files with ignores" at https://eslint.org/docs/latest/use/configure/configuration-files
+export default [
+    {
+        ignores: ["**/.eslintrc.js", "admin/words.js", "echarts.js"]
+    }
+];
